max_supply change int to long

+++ app/src/main/java/com/korugan/expansetrackerapp/data/remote/crypto/ticker/dto/CryptoTickerDto.kt
@@ -5,7 +5,7 @@ data class CryptoTickerDto(
     val first_data_at: String,
     val id: String,
     val last_updated: String,
-    val max_supply: Int,
+    val max_supply: Long,
     val name: String,
     val quotes: Quotes,
     val rank: Int,
